Read/Writer are now functional and the Client and Connection are both using them.  Writer is optionally threaded in order to attach to ObjectatEvents and receive serial updates.  I am not terribly happy with performance of the client for serial events, but I anticipated that issue.  I am thinking a non serial client is probably going to be the best, most scalable, method.  However, a serial feed could be useful for some applications.  CPU utilization jumped about 10% on my i7 2600K @ 4.2GHz when I started a client.  I had about 100 events per second beind synthetically generated at that time.  As I added threads with synthetic generators I didnt see the CPU utilization increase in parallel, so it may not be a huge deal.  It also may have to do with the polling cylce (10ms), that can probably be adjusted.
I also started to change my logging a bit to include the class name via
the this.getClass() method.  This seems to be more ideal when debugging
and I will attempt to modify all logging calls to follow this standard.

Need to add comments to most of the code still, I have been spending all
the time coding so far and not much at all leaving comments.

Added a client command to test suite that kicks off a client thread in
the background to consume resources, mainly for benchmarking.  I had it
printing out the data as it received it to ensure it was working, but I
disabled that during the stress testing done later as 100+ evens per
second scrolling across in XML docs would be unreadable and I want to
pull event stats as it runs.

Project is starting to come along, though I am concerned about
scalability while adding the client, ETC.  Also need to work on the
ability to tie multiple Objectats together via something like the client
serial messaging that currently exists.