Updated date format to use SimpleDateFormat and with the string format "yyyy-MM-dd HH:mm:ss zzz" for logger and event stats.
Updated disk writer to write to a temporary file first, delete the
primary and then rename the temp to the primary name.  I believe there
may be some caveats to implementation in Java across platforms per the
docs, but the targets for this application will likely work with the
current code.

Added debug logging around write and also increased frequency to 10
seconds between writes, since this is a single thread it will write,
wait 10 seconds and write again which may be fine.  This will likely be
exposed as a configuration option later on.

Testing seems to be going reltively well still.  Performance still seems
to be acceptable, I was able to push 600+ events per second on my
surface pro.  This is probably a very high number for the target
audience of this app and for the hardware.  Dedicated server hardware
will likely be able to attain much higher marks and the data wont all be
generated and processed in a single server necessarily.